fix for bug #507942 : bad macaddr of e1000 in Windows 2003 server with original Microsoft driver
Minor comment change (to last sent patch) and fixed WS issue of the patch.

The sequence of reading from eeprom is "offset by one" moved because of a false detection
of a clock cycle after an eeprom reset. Keeping the last clock value after a reset keeps it in sync.

Signed-off-by: Naphtali Sprei <nsprei@redhat.com>
Message-ID: <4A5EE609.3090104@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 507942
Acked-by: Juan Quintela <quintela@redhat.com>
Acked-by: Luiz Capitulino <lcapitulino@redhat.com>
Acked-by: Eduardo Habkost <ehabkost@redhat.com>