Set the iothread's eventfd/pipe descriptors to non-blocking
It replaces the previous fix of using select.
Fixes BZs: 511199, 507659 for sure, maybe 511031
Marcelo applied it to the upstream repository.

>From a2567128932437d286abc82ef8f79519f9992d74 Mon Sep 17 00:00:00 2001
From: Dor Laor <dor@redhat.com>
Date: Wed, 15 Jul 2009 17:53:16 +0300
Subject: [PATCH] Set the iothread's eventfd/pipe descriptors to non-blocking.

It fixes migration issue when the destination is loaded.

If the migration socket is full, we get EAGAIN for the write.
The set_fd_handler2 defers the write for later on. The function
tries to wake up the iothread by qemu_kvm_notify_work.
Since this happens in a loop, multiple times, the pipe that emulates eventfd
becomes full and we get a deadlock.

Mark McLoughlin suggested to remove spurious wake-up of the migration code
when we get EAGAIN and wait for the socket to become writeable. (+1)

Nevertheless, the pipe descriptors shouldn't be blocking and the reader can
also read several chunks in a time.

Signed-off-by: Dor Laor <dor@redhat.com>
Message-ID: <4A62E80E.6030401@redhat.com>
Obsoletes: <4A5DF084.3080305@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 511199
Bugzilla: 507659
Bugzilla-related: 511031
Acked-by: Juan Quintela <quintela@redhat.com>
Acked-by: Mark McLoughlin <markmc@redhat.com>
Acked-by: Glauber Costa <glommer@redhat.com>