ide calls register_savevm() three times for pci_ide_save/load
Hi

In this commit, it was changed the save/load version for ide to 3 in only 1 place,
but pci_ide_save/load is called from three places.  This patch fixes it.

commit ee49f80b360dd57acc61df1694fc7ea798ffa485
Author: Gleb Natapov <gleb@redhat.com>
Date:   Wed Jul 29 19:15:11 2009 +0300

As funny as it can get, the problem that we are fixing is that keyboard didn't worked
after migration.  If you really like strange bugs :)

Later, Juan.

bugzilla 514887

[ehabkost: more detailed description from Markus Armbruster below]

ide.c defines a couple of chips: ISA, CMD646, PIIX3, PIIX4,
PowerMac, MMIO, DSCM-1XXXX.

Gleb's commit changed pci_ide_save() and pci_ide_load() to put/get an
additional four bytes.  The two functions are used with CMD646, PIIX3
and PIIX3.  The commit bumped the savevm version_id only for CMD646, not
for PIIX3 and PIIX4.  Juan's patch bumps the other two.

Signed-off-by: Juan Quintela <quintela@redhat.com>
Message-Id: <1249059708-32209-1-git-send-email-quintela@redhat.com>
Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 514887
Acked-by: Eduardo Habkost <ehabkost@redhat.com>
Acked-by: Markus Armbruster <armbru@redhat.com>
Acked-by: Rik van Riel <riel@redhat.com>