Initialize PS2 keyboard / mouse state on reset
https://bugzilla.redhat.com/515275

This is a backport of qemu commit ef74679a810fe6858f625b9d52b68cc3fc61eb3d.

I don't know if this bug should be considered a RHEL-5.4 blocker. If not, then
this is a RHEL-5.5 patch submission.  :)

Log message from the patch added on Bugzilla:

    Initialize PS2 keyboard / mouse state on reset

    Currently only common PS2 state is initialized, leaving keyboard and
    mouse specific state to contain stale values.

    Signed-off-by: Dinesh Subhraveti <dineshs@us.ibm.com>
    Signed-off-by: Anthony Liguori <aliguori@us.ibm.com>
    Signed-off-by: Ram Pai <linuxram@us.ibm.com>

Signed-off-by: Eduardo Habkost <ehabkost@redhat.com>
Bugzilla: 517855
RH-Upstream-status: backport-from-upstream
Message-Id: <1249318832-29302-1-git-send-email-ehabkost@redhat.com>
Acked-by: Juan Quintela <quintela@redhat.com>
Acked-by: Markus Armbruster <armbru@redhat.com>
Acked-by: Dor Laor <dlaor@redhat.com>